+++ src/cross_bc.h
@@ -30,9 +30,9 @@ class BC : public QTLCross
                                            const Rcpp::IntegerVector& cross_info);
     const int ngen(const bool is_x_chr);
 
-    const double nrec(const int gen_left, const int gen_right,
-                      const bool is_x_chr, const bool is_female,
-                      const Rcpp::IntegerVector& cross_info);
+    const int nrec(const int gen_left, const int gen_right,
+                   const bool is_x_chr, const bool is_female,
+                   const Rcpp::IntegerVector& cross_info);
 
     const bool check_is_female_vector(const Rcpp::LogicalVector& is_female, const bool any_x_chr);
